feat: create all tables refactor: move each class in separated file
