+++ .gitignore
@@ -1,3 +1,6 @@
 env/
 .idea
 .env
+models/__pycache__/
+controllers/__pycache__/
+__pycache__/
