+++ models/base.py
@@ -1,0 +1,4 @@
+from sqlalchemy.orm import declarative_base
+
+
+Base = declarative_base()
